requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings>=2.0.0
google-genai>=1.30.0
chromadb==0.4.22
numpy==1.26.4
sentence-transformers>=2.2.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
aiofiles==23.2.1
pytz==2023.3
setuptools>=65.0.0
torch>=2.0.0
transformers>=4.30.0
